ofRandom: split Engine and Distribution
